+++ CHANGELOG.md
@@ -17,3 +17,5 @@
  - display time passed from the beginning in the client
  - remove client from this repository
  - server send tree data to the client
+ - provide distinct list of buildings inserted in the tree
+ - detect next level of a certain building